requirements.txt

uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pyyaml>=6.0.1
openai>=1.0.0
anthropic>=0.25.0
python-multipart>=0.0.6
httpx>=0.25.0
python-json-logger>=2.0.7
structlog>=23.2.0
python-dotenv>=1.0.0
litellm>=1.0.0
fastapi==0.116.1
mangum==0.19.0
strands-agents